refactor: put LineItemsTable into its own component
